+++ pages/api/auth/[...nextauth].js
@@ -26,27 +26,22 @@ export const authOptions = {
         }
       },
     }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-    }),
-    GitHubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET,
-    }),
+    process.env.VERCEL_ENV !== "preview" &&
+      GoogleProvider({
+        clientId: process.env.GOOGLE_CLIENT_ID,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+      }),
+    process.env.VERCEL_ENV !== "preview" &&
+      GitHubProvider({
+        clientId: process.env.GITHUB_CLIENT_ID,
+        clientSecret: process.env.GITHUB_CLIENT_SECRET,
+      }),
   ],
   callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.email = user.email;
-        token.name = user.name;
-        token.initial = user.name?.charAt(0);
-      }
-      return token;
-    },
     async session({ session, token }) {
       session.user.email = token.email;
       session.user.name = token.name;
+      session.user.id = token.sub;
       return session;
     },
   },
